pyproject.toml
[project]
name = "odoo-sales"
version = "0.1.0"
description = "Small Python library to fetch Odoo sales data by date range."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.0",
  "fastapi[standard]>=0.115.0",
  "authlib>=1.6.8",
  "httpx>=0.28.1",
  "itsdangerous>=2.2.0",
  "google-cloud-firestore>=2.23.0",
]

[project.scripts]
odoo-sales = "odoo_sales.cli:main"
odoo-sales-web = "odoo_sales.web:main"
odoo-sales-users = "odoo_sales.users:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
